refactor(registration): extract waitlist admin operations
- Extract handleClearWaitlist -> handleClearWaitlistOp
- Extract handleRemoveFromWaitlist -> handleRemoveFromWaitlistOp
- Create src/registration/handlers/adminOperations.js
- Thin wrappers preserve original function names in App.jsx
